chore: move serde impls to module (#62)
