+++ Cargo.toml
@@ -5,7 +5,7 @@ version = "1.3.0"
 authors = ["Hyunsik Choi <hyunsik.choi@gmail.com>", "MrCroxx <mrcroxx@outlook.com>"]
 keywords = ["byte", "byte-size", "utility", "human-readable", "format"]
 categories = ["development-tools", "filesystem"]
-repository = "https://github.com/hyunsik/bytesize"
+repository = "https://github.com/bytesize-rs/bytesize"
 license = "Apache-2.0"
 edition = "2021"
 rust-version = "1.65"
